Parameter types and return type change to double from integer

+++ Calculator.java
@@ -2,7 +2,7 @@ public class Claculator{
     public static void main(String[] args){
 
     }
-    int addition(int x, int y){
+    double addition(double x, double y){
         return x+y;
     }
 }
